refactor: 수신자(Receiver) 관련 패키지 구조 재구성 및 파일 위치 최적화
- 수신자 관련 컴포넌트, 모델, 뷰모델을 기능별 도메인 패키지(`home`, `list`, `summary`, `afternotemain`, `trigger`, `playlist`)로 이동
- `ReceiverAfternoteListEntry`를 `ReceiverAfternoteHomeEntry`로 변경하고 `home` 패키지로 이동
- `ReceiverAfternoteHomeViewModel`, `ReceiverAfternoteHomeEvent`, `ReceiverAfternoteHomeUiState`를 `home` 패키지로 통합
- `ReceiverAfternotesListViewModel`, `ReceiverAfternotesListEvent` 등 리스트 관련 파일을 `list` 패키지로 이동
- `ReceiverDownloadAllViewModel`, `ReceiverSummaryUiState` 및 관련 UI 컴포넌트(`HeroCard`, `TopHeader` 등)를 `summary` 패키지로 이동
- `ReceiverMemorialPlaylistViewModel` 및 관련 상태/이벤트를 `playlist` 패키지로 이동
- 패키지 구조 변경에 따른 프로젝트 내 임포트 경로 업데이트 및 정리
