+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/receiver/afternotemain/ReceiverAfterNoteMainScreen.kt
@@ -1,4 +1,4 @@
-package com.afternote.feature.afternote.presentation.receiver.screen.afternote
+package com.afternote.feature.afternote.presentation.receiver.afternotemain
 
 import android.content.Intent
 import android.content.pm.PackageManager
