+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/receiver/home/ReceiverAfternoteHomeEntry.kt
@@ -1,4 +1,4 @@
-package com.afternote.feature.afternote.presentation.receiver.navgraph
+package com.afternote.feature.afternote.presentation.receiver.home
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -7,8 +7,6 @@ import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.afternote.core.ui.scaffold.bottombar.BottomNavTab
 import com.afternote.feature.afternote.presentation.author.home.AfternoteHomeScreen
-import com.afternote.feature.afternote.presentation.receiver.model.ReceiverAfternoteHomeEvent
-import com.afternote.feature.afternote.presentation.receiver.viewmodel.ReceiverAfternoteHomeViewModel
 import com.afternote.feature.afternote.presentation.shared.body.infinite.AfternoteBodyUiState
 
 data class ReceiverAfternoteHomeEntryActions(
