+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/receiver/home/ReceiverAfternoteHomeUiState.kt
@@ -1,4 +1,4 @@
-package com.afternote.feature.afternote.presentation.receiver.model.uistate
+package com.afternote.feature.afternote.presentation.receiver.home
 import com.afternote.core.ui.scaffold.bottombar.BottomNavTab
 import com.afternote.feature.afternote.presentation.shared.AfternoteCategory
 import com.afternote.feature.afternote.presentation.shared.body.infinite.content.list.item.ListItemUiModel
