+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/receiver/list/ReceiverAfternotesListViewModel.kt
@@ -1,11 +1,8 @@
-package com.afternote.feature.afternote.presentation.receiver.viewmodel
+package com.afternote.feature.afternote.presentation.receiver.list
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.afternote.feature.afternote.domain.repository.ReceiverRepository
-import com.afternote.feature.afternote.presentation.receiver.model.ReceiverAfternotesListEvent
-import com.afternote.feature.afternote.presentation.receiver.model.uistate.ReceivedAfternoteListItemUi
-import com.afternote.feature.afternote.presentation.receiver.model.uistate.ReceiverAfternotesListUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
