+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/receiver/summary/HeroCard.kt
@@ -1,4 +1,4 @@
-package com.afternote.feature.afternote.presentation.receiver.component
+package com.afternote.feature.afternote.presentation.receiver.summary
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
