+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/receiver/summary/ReceiverDownloadAllViewModel.kt
@@ -1,9 +1,8 @@
-package com.afternote.feature.afternote.presentation.receiver.viewmodel
+package com.afternote.feature.afternote.presentation.receiver.summary
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.afternote.feature.afternote.domain.repository.ReceiverRepository
-import com.afternote.feature.afternote.presentation.receiver.model.uistate.ReceiverDownloadAllUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
