+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/receiver/trigger/ReceiverAfternoteTriggerEvent.kt
@@ -1,4 +1,4 @@
-package com.afternote.feature.afternote.presentation.receiver.model
+package com.afternote.feature.afternote.presentation.receiver.trigger
 
 /**
  * 수신자 트리거 화면에서 발생하는 사용자 이벤트.
